Move src/config.ts to src/lib/
