+++ src/background.ts
@@ -8,7 +8,7 @@ import * as messaging from "./messaging"
 import * as excmds from "./.excmds_background.generated"
 import * as commandline_background from "@src/background/commandline_background"
 import * as convert from "./convert"
-import * as config from "./config"
+import * as config from "@src/lib/config"
 import * as dom from "./dom"
 import * as download_background from "./download_background"
 import * as itertools from "./itertools"
