+++ src/commandline_frame.ts
@@ -10,7 +10,7 @@ import { ExcmdCompletionSource } from "./completions/Excmd"
 import { HistoryCompletionSource } from "./completions/History"
 import { SettingsCompletionSource } from "./completions/Settings"
 import * as Messaging from "./messaging"
-import * as Config from "./config"
+import * as Config from "@src/lib/config"
 import * as SELF from "./commandline_frame"
 import "./number.clamp"
 import state from "./state"
