+++ src/completions.ts
@@ -13,7 +13,7 @@ Concrete completion classes have been moved to src/completions/.
 import * as Fuse from "fuse.js"
 import { enumerate } from "./itertools"
 import { toNumber } from "./convert"
-import * as config from "./config"
+import * as config from "@src/lib/config"
 import * as aliases from "@src/lib/aliases"
 
 export const DEFAULT_FAVICON = browser.extension.getURL(
