+++ src/completions/Settings.ts
@@ -1,5 +1,5 @@
 import * as Completions from "../completions"
-import * as config from "../config"
+import * as config from "@src/lib/config"
 import { browserBg } from "../lib/webext"
 import * as metadata from "../.metadata.generated"
 import { typeToString } from "../metadata"
