+++ src/excmds.ts
@@ -97,7 +97,7 @@ import * as Container from "./lib/containers"
 import state from "./state"
 import { contentState, ModeName } from "./state_content"
 import * as UrlUtil from "./url_util"
-import * as config from "./config"
+import * as config from "@src/lib/config"
 import * as aliases from "@src/lib/aliases"
 import * as Logging from "@src/lib/logging"
 /** @hidden */
