+++ src/finding.ts
@@ -14,7 +14,7 @@ import * as DOM from "./dom"
 import { hasModifiers } from "./keyseq"
 import { contentState } from "./state_content"
 import { messageActiveTab, message } from "./messaging"
-import * as config from "./config"
+import * as config from "@src/lib/config"
 import Logger from "@src/lib/logging"
 import Mark from "mark.js"
 const logger = new Logger("finding")
