+++ src/native_background.ts
@@ -3,7 +3,7 @@
  */
 
 import * as semverCompare from "semver-compare"
-import * as config from "./config"
+import * as config from "@src/lib/config"
 import { browserBg } from "./lib/webext"
 
 import Logger from "@src/lib/logging"
