+++ src/parsers/exmode.ts
@@ -2,7 +2,7 @@
 
 import * as ExCmds from "../.excmds_background.generated"
 import * as convert from "../convert"
-import * as Config from "../config"
+import * as Config from "@src/lib/config"
 import * as aliases from "@src/lib/aliases"
 import * as Logging from "@src/lib/logging"
 import { enumerate, head, izip } from "../itertools"
