+++ src/requests.ts
@@ -1,7 +1,7 @@
 import * as Messaging from "./messaging"
 import * as Container from "./lib/containers"
 import * as UrlUtil from "./url_util"
-import * as config from "./config"
+import * as config from "@src/lib/config"
 import * as csp from "csp-serdes"
 import Logger from "@src/lib/logging"
