+++ src/styling.ts
@@ -1,5 +1,5 @@
 import { staticThemes } from "./.metadata.generated"
-import * as config from "./config"
+import * as config from "@src/lib/config"
 import * as Logging from "@src/lib/logging"
 
 const logger = new Logging.Logger("styling")
