+++ src/text_to_speech.ts
@@ -1,7 +1,7 @@
 /** Functions to deal with text to speech in Tridactyl
  */
 
-import * as Config from "./config"
+import * as Config from "@src/lib/config"
 
 /** Find the voice object for a voice name
  *
